refactor: split server.js monolith into route/service modules
server.js was 2187 lines mixing routing, auth, sort algorithms, CSV
parsing, and price-simulation math. Split into:

- middleware/auth.js: authenticateToken, requireAdmin, authLimiter
- utils/authHelpers.js: verifyPassword, generateSession
- utils/priceHelpers.js: resolveCardPrice, getSortedPositionForCard,
  rebalanceLocationPositions, getSimulatedPriceAt, isVintageSet
- routes/auth.js, shared.js, admin.js, collection.js, decks.js

server.js is now ~100 lines: app setup, middleware, router mounting,
static serving, and startup. No behavior changes — verified via curl
against every route family (auth, locations, collection, stats, decks,
admin, shared, export) plus CORS and rate-limit checks post-split.
