+++ backend/src/server.js
@@ -3,11 +3,15 @@ require('dotenv').config();
 const express = require('express');
 const cors = require('cors');
 const helmet = require('helmet');
-const rateLimit = require('express-rate-limit');
 const path = require('path');
-const crypto = require('crypto');
 const db = require('./db');
 const tcgApi = require('./tcgApi');
+
+const authRoutes = require('./routes/auth');
+const sharedRoutes = require('./routes/shared');
+const adminRoutes = require('./routes/admin');
+const collectionRoutes = require('./routes/collection');
+const decksRoutes = require('./routes/decks');
 
 const app = express();
 const PORT = process.env.PORT || 3001;
@@ -32,15 +36,6 @@ app.use(cors({
   }
 }));
 app.use(express.json());
-
-// Throttle auth endpoints to slow down credential-stuffing/brute-force attempts.
-const authLimiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  limit: 20,
-  standardHeaders: true,
-  legacyHeaders: false,
-  message: { error: 'Too many attempts. Please try again later.' }
-});
 
 // Initialize Database on startup
 db.initDb()
@@ -67,2094 +62,12 @@ db.initDb()
     console.error('Failed to initialize database:', err);
   });
 
-// --- AUTHENTICATION HELPERS & MIDDLEWARE ---
-
-function verifyPassword(password, storedHash) {
-  if (!storedHash) return false;
-  const parts = storedHash.split(':');
-
-  let iterations, salt, hash;
-  if (parts.length === 3) {
-    [iterations, salt, hash] = parts;
-    iterations = parseInt(iterations, 10);
-  } else if (parts.length === 2) {
-    // Legacy hashes created before the iteration count was stored per-hash.
-    iterations = 10000;
-    [salt, hash] = parts;
-  } else {
-    return false;
-  }
-
-  const storedBuf = Buffer.from(hash, 'hex');
-  const verifyBuf = crypto.pbkdf2Sync(password, salt, iterations, 64, 'sha512');
-  if (storedBuf.length !== verifyBuf.length) return false;
-  return crypto.timingSafeEqual(storedBuf, verifyBuf);
-}
-
-function resolveCardPrice(card) {
-  if (!card) return 0;
-  if (card.printing === 'Holofoil' && card.price_holofoil !== null && card.price_holofoil > 0) {
-    return card.price_holofoil;
-  }
-  if (card.printing === 'Reverse Holofoil' && card.price_reverse_holofoil !== null && card.price_reverse_holofoil > 0) {
-    return card.price_reverse_holofoil;
-  }
-  if (card.printing === 'Normal' && card.price_normal !== null && card.price_normal > 0) {
-    return card.price_normal;
-  }
-  return card.price_trend || 0;
-}
-
-async function generateSession(userId) {
-  const token = crypto.randomBytes(32).toString('hex');
-  const expiresAt = new Date();
-  expiresAt.setDate(expiresAt.getDate() + 7); // 7 days expiration
-  const expiresAtStr = expiresAt.toISOString().slice(0, 19).replace('T', ' '); // YYYY-MM-DD HH:MM:SS
-  await db.run(`
-    INSERT INTO sessions (token, user_id, expires_at)
-    VALUES (?, ?, ?)
-  `, [token, userId, expiresAtStr]);
-  return token;
-}
-
-async function rebalanceLocationPositions(db, locationId, userId) {
-  if (!locationId) return;
-  const cards = await db.all(`SELECT id FROM collection WHERE location_id = ? AND user_id = ? ORDER BY position ASC`, [locationId, userId]);
-  for (let i = 0; i < cards.length; i++) {
-    const cleanPos = (i + 1) * 1000;
-    await db.run(`UPDATE collection SET position = ? WHERE id = ?`, [cleanPos, cards[i].id]);
-  }
-}
-
-async function getSortedPositionForCard(db, locationId, userId, cardMetadata) {
-  const loc = await db.get(`SELECT type, sort_order, foil_sorting FROM locations WHERE id = ? AND user_id = ?`, [locationId, userId]);
-  if (!loc || loc.sort_order === 'custom') {
-    const maxRow = await db.get(`SELECT MAX(position) as maxPos FROM collection WHERE location_id = ? AND user_id = ?`, [locationId, userId]);
-    return maxRow && maxRow.maxPos !== null ? maxRow.maxPos + 1000 : 1000;
-  }
-
-  const sortOrder = loc.sort_order;
-
-  const query = `
-    SELECT c.id as entry_id, c.position, c.printing, cc.name, cc.supertype, cc.types, cc.rarity, cc.set_name, cc.number, cc.price_trend, cc.price_normal, cc.price_holofoil, cc.price_reverse_holofoil
-    FROM collection c
-    JOIN card_cache cc ON c.card_id = cc.id
-    WHERE c.location_id = ? AND c.user_id = ?
-  `;
-  const existing = await db.all(query, [locationId, userId]);
-
-  existing.forEach(c => {
-    try {
-      c.types = JSON.parse(c.types || '[]');
-    } catch {
-      c.types = [];
-    }
-    c.price_trend = resolveCardPrice(c);
-  });
-
-  const newCard = {
-    entry_id: -1,
-    printing: cardMetadata.printing || 'Normal',
-    name: cardMetadata.name || '',
-    supertype: cardMetadata.supertype || '',
-    types: cardMetadata.types || [],
-    rarity: cardMetadata.rarity || '',
-    set_name: cardMetadata.set_name || '',
-    number: cardMetadata.number || '0',
-    price_trend: resolveCardPrice(cardMetadata)
-  };
-  existing.push(newCard);
-
-  const POKEMON_TYPE_ORDER = {
-    'Grass': 1, 'Fire': 2, 'Water': 3, 'Lightning': 4, 'Psychic': 5,
-    'Fighting': 6, 'Darkness': 7, 'Metal': 8, 'Dragon': 9, 'Colorless': 10, 'Trainer': 11, 'Energy': 12
-  };
-  
-  const PRINTING_ORDER_NORMALS_FIRST = {
-    'Normal': 1,
-    'Reverse Holofoil': 2,
-    'Holofoil': 3,
-    '1st Edition': 4,
-    'Promo': 5
-  };
-  
-  const PRINTING_ORDER_FOILS_FIRST = {
-    'Reverse Holofoil': 1,
-    'Holofoil': 2,
-    'Normal': 3,
-    '1st Edition': 4,
-    'Promo': 5
-  };
-
-  const isFoilsFirst = loc && loc.foil_sorting === 'foils_first';
-  const PRINTING_ORDER = isFoilsFirst ? PRINTING_ORDER_FOILS_FIRST : PRINTING_ORDER_NORMALS_FIRST;
-
-  if (sortOrder === 'name-asc') {
-    existing.sort((a, b) => a.name.localeCompare(b.name));
-  } else if (sortOrder === 'price-desc') {
-    existing.sort((a, b) => (b.price_trend || 0) - (a.price_trend || 0));
-  } else if (sortOrder === 'set-number') {
-    existing.sort((a, b) => {
-      const cmpSet = (a.set_name || '').localeCompare(b.set_name || '');
-      if (cmpSet !== 0) return cmpSet;
-      const numA = parseInt(a.number || '0', 10) || 0;
-      const numB = parseInt(b.number || '0', 10) || 0;
-      if (numA !== numB) return numA - numB;
-      return (a.number || '').localeCompare(b.number || '');
-    });
-  } else if (sortOrder === 'set-number-printing') {
-    existing.sort((a, b) => {
-      const setA = a.set_name || '';
-      const setB = b.set_name || '';
-      const cmpSet = setA.localeCompare(setB);
-      if (cmpSet !== 0) return cmpSet;
-
-      const printA = PRINTING_ORDER[a.printing] || 10;
-      const printB = PRINTING_ORDER[b.printing] || 10;
-      if (printA !== printB) return printA - printB;
-
-      const numA = parseInt(a.number || '0', 10) || 0;
-      const numB = parseInt(b.number || '0', 10) || 0;
-      if (numA !== numB) return numA - numB;
-
-      const cmpNum = (a.number || '').localeCompare(b.number || '');
-      if (cmpNum !== 0) return cmpNum;
-
-      return a.name.localeCompare(b.name);
-    });
-  } else if (sortOrder === 'type-name') {
-    existing.sort((a, b) => {
-      const typeA = (a.types && a.types[0]) || 'Unknown';
-      const typeB = (b.types && b.types[0]) || 'Unknown';
-      const orderA = POKEMON_TYPE_ORDER[typeA] || 50;
-      const orderB = POKEMON_TYPE_ORDER[typeB] || 50;
-      if (orderA !== orderB) return orderA - orderB;
-      return a.name.localeCompare(b.name);
-    });
-  }
-
-  const targetIndex = existing.findIndex(c => c.entry_id === -1);
-  if (targetIndex === -1) return 1000;
-
-  if (existing.length === 1) return 1000;
-  if (targetIndex === 0) {
-    return existing[1].position / 2;
-  }
-  if (targetIndex === existing.length - 1) {
-    return existing[targetIndex - 1].position + 1000;
-  }
-  return (existing[targetIndex - 1].position + existing[targetIndex + 1].position) / 2;
-}
-
-async function authenticateToken(req, res, next) {
-  let token = null;
-  const authHeader = req.headers['authorization'];
-  if (authHeader && authHeader.startsWith('Bearer ')) {
-    token = authHeader.split(' ')[1];
-  }
-
-  if (!token) {
-    return res.status(401).json({ error: 'Access denied. No token provided.' });
-  }
-
-  try {
-    const session = await db.get(`
-      SELECT s.user_id, u.username, u.role, u.share_token, u.share_enabled, u.tcg_api_key
-      FROM sessions s
-      JOIN users u ON s.user_id = u.id
-      WHERE s.token = ? AND s.expires_at > DATETIME('now')
-    `, [token]);
-
-    if (!session) {
-      return res.status(401).json({ error: 'Invalid or expired session token.' });
-    }
-
-    req.user = {
-      id: session.user_id,
-      username: session.username,
-      role: session.role,
-      share_token: session.share_token,
-      share_enabled: session.share_enabled,
-      tcg_api_key: session.tcg_api_key || ''
-    };
-    next();
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Authentication error' });
-  }
-}
-
 // --- API ROUTES ---
-
-// --- USER AUTHENTICATION & PROFILE ENDPOINTS ---
-
-// Register a new user
-app.post('/api/auth/register', authLimiter, async (req, res) => {
-  const { username, password } = req.body;
-  if (!username || !password) {
-    return res.status(400).json({ error: 'Username and password are required' });
-  }
-
-  const cleanUsername = username.trim().toLowerCase();
-  if (cleanUsername.length < 3) {
-    return res.status(400).json({ error: 'Username must be at least 3 characters' });
-  }
-  if (password.length < 8) {
-    return res.status(400).json({ error: 'Password must be at least 8 characters' });
-  }
-
-  try {
-    const existingUser = await db.get(`SELECT id FROM users WHERE username = ?`, [cleanUsername]);
-    if (existingUser) {
-      return res.status(400).json({ error: 'Username is already taken' });
-    }
-
-    const passwordHash = db.hashPassword(password);
-    const shareToken = crypto.randomBytes(16).toString('hex');
-
-    const result = await db.run(`
-      INSERT INTO users (username, password_hash, role, share_token, share_enabled)
-      VALUES (?, ?, ?, ?, ?)
-    `, [cleanUsername, passwordHash, 'member', shareToken, 0]);
-
-    const token = await generateSession(result.lastID);
-
-    res.status(201).json({
-      message: 'Registration successful',
-      token,
-      user: {
-        username: cleanUsername,
-        role: 'member',
-        share_token: shareToken,
-        share_enabled: 0
-      }
-    });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to register' });
-  }
-});
-
-// Login user
-app.post('/api/auth/login', authLimiter, async (req, res) => {
-  const { username, password } = req.body;
-  if (!username || !password) {
-    return res.status(400).json({ error: 'Username and password are required' });
-  }
-
-  const cleanUsername = username.trim().toLowerCase();
-
-  try {
-    const user = await db.get(`SELECT * FROM users WHERE username = ?`, [cleanUsername]);
-    if (!user || !verifyPassword(password, user.password_hash)) {
-      return res.status(401).json({ error: 'Invalid username or password' });
-    }
-
-    const token = await generateSession(user.id);
-
-    res.json({
-      message: 'Login successful',
-      token,
-      user: {
-        username: user.username,
-        role: user.role,
-        share_token: user.share_token,
-        share_enabled: user.share_enabled,
-        tcg_api_key: user.tcg_api_key || ''
-      }
-    });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Login failed' });
-  }
-});
-
-// Logout user
-app.post('/api/auth/logout', authenticateToken, async (req, res) => {
-  const authHeader = req.headers['authorization'];
-  const token = authHeader && authHeader.split(' ')[1];
-  
-  try {
-    if (token) {
-      await db.run(`DELETE FROM sessions WHERE token = ?`, [token]);
-    }
-    res.json({ message: 'Logged out successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Logout failed' });
-  }
-});
-
-// Get current user profile
-app.get('/api/auth/me', authenticateToken, (req, res) => {
-  res.json({ user: req.user });
-});
-
-// Update settings (password, sharing)
-app.put('/api/auth/settings', authenticateToken, async (req, res) => {
-  const { current_password, password, share_enabled, regenerate_share_token, tcg_api_key } = req.body;
-
-  try {
-    if (password !== undefined) {
-      if (password.length < 8) {
-        return res.status(400).json({ error: 'Password must be at least 8 characters' });
-      }
-      const currentUser = await db.get(`SELECT password_hash FROM users WHERE id = ?`, [req.user.id]);
-      if (!current_password || !verifyPassword(current_password, currentUser.password_hash)) {
-        return res.status(401).json({ error: 'Current password is incorrect' });
-      }
-      const newHash = db.hashPassword(password);
-      await db.run(`UPDATE users SET password_hash = ? WHERE id = ?`, [newHash, req.user.id]);
-    }
-
-    if (share_enabled !== undefined) {
-      await db.run(`UPDATE users SET share_enabled = ? WHERE id = ?`, [share_enabled ? 1 : 0, req.user.id]);
-    }
-
-    if (tcg_api_key !== undefined) {
-      await db.run(`UPDATE users SET tcg_api_key = ? WHERE id = ?`, [tcg_api_key.trim(), req.user.id]);
-    }
-
-    let newShareToken = req.user.share_token;
-    if (regenerate_share_token) {
-      newShareToken = crypto.randomBytes(16).toString('hex');
-      await db.run(`UPDATE users SET share_token = ? WHERE id = ?`, [newShareToken, req.user.id]);
-    }
-
-    // Retrieve updated info
-    const updatedUser = await db.get(`SELECT username, role, share_token, share_enabled, tcg_api_key FROM users WHERE id = ?`, [req.user.id]);
-    res.json({
-      message: 'Settings updated successfully',
-      user: {
-        username: updatedUser.username,
-        role: updatedUser.role,
-        share_token: updatedUser.share_token,
-        share_enabled: updatedUser.share_enabled,
-        tcg_api_key: updatedUser.tcg_api_key || ''
-      }
-    });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to update settings' });
-  }
-});
-
-// --- PUBLIC SHARING ENDPOINT ---
-
-// Retrieve a shared collection by share token
-app.get('/api/shared/:share_token', async (req, res) => {
-  const { share_token } = req.params;
-  const listType = req.query.list || 'collection';
-
-  try {
-    const owner = await db.get(`SELECT id, username, share_enabled FROM users WHERE share_token = ?`, [share_token]);
-    if (!owner || owner.share_enabled === 0) {
-      return res.status(404).json({ error: 'This card collection is private or does not exist.' });
-    }
-
-    let filterSql = `WHERE c.user_id = ?`;
-    let filterParams = [owner.id];
-
-    if (listType === 'wishlist') {
-      filterSql += ` AND c.list_type = 'wishlist'`;
-    } else if (listType === 'trade') {
-      filterSql += ` AND c.is_trade = 1 AND c.list_type = 'collection'`;
-    } else {
-      filterSql += ` AND c.list_type = 'collection'`;
-    }
-
-    // Retrieve their collection without private fields (locations, purchase price, ROI)
-    const query = `
-      SELECT 
-        c.id as entry_id,
-        c.card_id,
-        c.quantity,
-        c.condition,
-        c.printing,
-        c.language,
-        c.added_at,
-        c.is_trade,
-        c.list_type,
-        cc.name,
-        cc.supertype,
-        cc.subtypes,
-        cc.types,
-        cc.rarity,
-        cc.set_id,
-        cc.set_name,
-        cc.number,
-        cc.image_url,
-        cc.price_trend,
-        cc.price_normal,
-        cc.price_holofoil,
-        cc.price_reverse_holofoil
-      FROM collection c
-      JOIN card_cache cc ON c.card_id = cc.id
-      ${filterSql}
-      ORDER BY c.added_at DESC
-    `;
-    const rows = await db.all(query, filterParams);
-
-    const formatted = rows.map(row => ({
-      ...row,
-      price_trend: resolveCardPrice(row),
-      subtypes: JSON.parse(row.subtypes || '[]'),
-      types: JSON.parse(row.types || '[]'),
-    }));
-
-    // Calculate public stats
-    let totalCards = 0;
-    let uniqueCards = formatted.length;
-    let totalValue = 0;
-
-    const typeCounts = {};
-    const rarityCounts = {};
-    const setCounts = {};
-
-    formatted.forEach(row => {
-      const qty = row.quantity || 1;
-      const price = row.price_trend || 0;
-
-      totalCards += qty;
-      totalValue += qty * price;
-
-      row.types.forEach(t => {
-        typeCounts[t] = (typeCounts[t] || 0) + qty;
-      });
-      if (row.types.length === 0) {
-        typeCounts['Colorless'] = (typeCounts['Colorless'] || 0) + qty;
-      }
-
-      const rarity = row.rarity || 'Unknown';
-      rarityCounts[rarity] = (rarityCounts[rarity] || 0) + qty;
-
-      if (!setCounts[row.set_id]) {
-        setCounts[row.set_id] = { name: row.set_name, count: 0, value: 0 };
-      }
-      setCounts[row.set_id].count += qty;
-      setCounts[row.set_id].value += qty * price;
-    });
-
-    res.json({
-      owner: owner.username,
-      collection: formatted,
-      stats: {
-        summary: {
-          totalCards,
-          uniqueCards,
-          totalValue: parseFloat(totalValue.toFixed(2))
-        },
-        types: Object.keys(typeCounts).map(name => ({ name, value: typeCounts[name] })),
-        rarities: Object.keys(rarityCounts).map(name => ({ name, value: rarityCounts[name] })),
-        sets: Object.keys(setCounts).map(id => ({ 
-          id, 
-          name: setCounts[id].name, 
-          count: setCounts[id].count, 
-          value: parseFloat(setCounts[id].value.toFixed(2)) 
-        })).sort((a,b) => b.value - a.value).slice(0, 8)
-      }
-    });
-
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to retrieve shared collection' });
-  }
-});
-
-// --- ADMIN ENDPOINTS (Only accessible by users with role 'admin') ---
-
-// Helper middleware to restrict to admin
-function requireAdmin(req, res, next) {
-  if (req.user && req.user.role === 'admin') {
-    next();
-  } else {
-    res.status(403).json({ error: 'Access denied. Administrator privileges required.' });
-  }
-}
-
-// Get all users with their statistics
-app.get('/api/admin/users', authenticateToken, requireAdmin, async (req, res) => {
-  try {
-    const users = await db.all(`
-      SELECT id, username, role, share_enabled, created_at
-      FROM users
-      ORDER BY username ASC
-    `);
-
-    // Fetch stats for each user
-    const usersWithStats = [];
-    for (const u of users) {
-      const stats = await db.get(`
-        SELECT COUNT(c.id) as unique_cards, SUM(c.quantity) as total_cards, 
-          SUM(c.quantity * CASE 
-            WHEN c.printing = 'Holofoil' AND cc.price_holofoil IS NOT NULL AND cc.price_holofoil > 0 THEN cc.price_holofoil
-            WHEN c.printing = 'Reverse Holofoil' AND cc.price_reverse_holofoil IS NOT NULL AND cc.price_reverse_holofoil > 0 THEN cc.price_reverse_holofoil
-            WHEN c.printing = 'Normal' AND cc.price_normal IS NOT NULL AND cc.price_normal > 0 THEN cc.price_normal
-            ELSE cc.price_trend
-          END) as total_value
-        FROM collection c
-        JOIN card_cache cc ON c.card_id = cc.id
-        WHERE c.user_id = ?
-      `, [u.id]);
-
-      usersWithStats.push({
-        ...u,
-        total_cards: stats.total_cards || 0,
-        unique_cards: stats.unique_cards || 0,
-        total_value: parseFloat((stats.total_value || 0).toFixed(2))
-      });
-    }
-
-    res.json(usersWithStats);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to retrieve users list' });
-  }
-});
-
-// Create a new user from Admin Panel
-app.post('/api/admin/users', authenticateToken, requireAdmin, async (req, res) => {
-  const { username, password, role = 'member' } = req.body;
-  if (!username || !password) {
-    return res.status(400).json({ error: 'Username and password are required' });
-  }
-
-  const cleanUsername = username.trim().toLowerCase();
-  if (cleanUsername.length < 3) {
-    return res.status(400).json({ error: 'Username must be at least 3 characters' });
-  }
-  if (password.length < 8) {
-    return res.status(400).json({ error: 'Password must be at least 8 characters' });
-  }
-  if (role !== 'member' && role !== 'admin') {
-    return res.status(400).json({ error: 'Invalid role specification' });
-  }
-
-  try {
-    const existingUser = await db.get(`SELECT id FROM users WHERE username = ?`, [cleanUsername]);
-    if (existingUser) {
-      return res.status(400).json({ error: 'Username is already taken' });
-    }
-
-    const passwordHash = db.hashPassword(password);
-    const shareToken = crypto.randomBytes(16).toString('hex');
-
-    await db.run(`
-      INSERT INTO users (username, password_hash, role, share_token, share_enabled)
-      VALUES (?, ?, ?, ?, ?)
-    `, [cleanUsername, passwordHash, role, shareToken, 0]);
-
-    res.status(201).json({ message: `User "${cleanUsername}" created successfully.` });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to create user' });
-  }
-});
-
-// Update a user (Change password or Role) from Admin Panel
-app.put('/api/admin/users/:id', authenticateToken, requireAdmin, async (req, res) => {
-  const { id } = req.params;
-  const { password, role } = req.body;
-
-  try {
-    const targetUser = await db.get(`SELECT id, username, role FROM users WHERE id = ?`, [id]);
-    if (!targetUser) {
-      return res.status(404).json({ error: 'User not found' });
-    }
-
-    if (password !== undefined) {
-      if (password.length < 8) {
-        return res.status(400).json({ error: 'Password must be at least 8 characters' });
-      }
-      const newHash = db.hashPassword(password);
-      await db.run(`UPDATE users SET password_hash = ? WHERE id = ?`, [newHash, id]);
-    }
-
-    if (role !== undefined) {
-      if (role !== 'member' && role !== 'admin') {
-        return res.status(400).json({ error: 'Invalid role' });
-      }
-      // Block admin demoting themselves
-      if (parseInt(id, 10) === req.user.id && role !== 'admin') {
-        return res.status(400).json({ error: 'You cannot demote yourself from Administrator role.' });
-      }
-      await db.run(`UPDATE users SET role = ? WHERE id = ?`, [role, id]);
-    }
-
-    res.json({ message: 'User updated successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to update user' });
-  }
-});
-
-// Delete user from Admin Panel
-app.delete('/api/admin/users/:id', authenticateToken, requireAdmin, async (req, res) => {
-  const { id } = req.params;
-
-  if (parseInt(id, 10) === req.user.id) {
-    return res.status(400).json({ error: 'You cannot delete your own Administrator account.' });
-  }
-
-  try {
-    const targetUser = await db.get(`SELECT id, username FROM users WHERE id = ?`, [id]);
-    if (!targetUser) {
-      return res.status(404).json({ error: 'User not found' });
-    }
-
-    await db.run(`DELETE FROM sessions WHERE user_id = ?`, [id]);
-    await db.run(`DELETE FROM collection WHERE user_id = ?`, [id]);
-    await db.run(`DELETE FROM locations WHERE user_id = ?`, [id]);
-    await db.run(`DELETE FROM users WHERE id = ?`, [id]);
-
-    res.json({ message: `User "${targetUser.username}" and all their card collections/locations have been permanently deleted.` });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to delete user' });
-  }
-});
-
-// Generate a random collection of various cards for admins database so we can test.
-// Dev/test-data helper only — unreachable once NODE_ENV=production.
-app.post('/api/admin/seed-cards', authenticateToken, requireAdmin, async (req, res) => {
-  if (process.env.NODE_ENV === 'production') {
-    return res.status(403).json({ error: 'Not available in production' });
-  }
-  try {
-    // 1. Get or create Binder and Box locations for admin
-    let binder = await db.get(`SELECT id, page_style FROM locations WHERE user_id = ? AND type = 'Binder' LIMIT 1`, [req.user.id]);
-    if (!binder) {
-      const result = await db.run(`
-        INSERT INTO locations (name, type, description, sort_order, max_pages, page_style, user_id)
-        VALUES (?, ?, ?, ?, ?, ?, ?)
-      `, ['Binder Seed Box', 'Binder', 'Seeded binder', 'custom', 10, '3x3', req.user.id]);
-      binder = { id: result.lastID, page_style: '3x3' };
-    }
-
-    let box = await db.get(`SELECT id FROM locations WHERE user_id = ? AND type = 'Box' LIMIT 1`, [req.user.id]);
-    if (!box) {
-      const result = await db.run(`
-        INSERT INTO locations (name, type, description, sort_order, max_rows, max_capacity, user_id)
-        VALUES (?, ?, ?, ?, ?, ?, ?)
-      `, ['Box Seed Box', 'Box', 'Seeded bulk box', 'custom', 3, 200, req.user.id]);
-      box = { id: result.lastID };
-    }
-
-    const MOCK_POOL = [
-      { id: 'base1-58', name: 'Pikachu', set_id: 'base1', set_name: 'Base Set', number: '58', rarity: 'Common', supertype: 'Pokémon', types: '["Lightning"]', img: 'https://images.pokemontcg.io/base1/58_hier.png' },
-      { id: 'base1-4', name: 'Charizard', set_id: 'base1', set_name: 'Base Set', number: '4', rarity: 'Rare Holo', supertype: 'Pokémon', types: '["Fire"]', img: 'https://images.pokemontcg.io/base1/4_hier.png' },
-      { id: 'base1-2', name: 'Blastoise', set_id: 'base1', set_name: 'Base Set', number: '2', rarity: 'Rare Holo', supertype: 'Pokémon', types: '["Water"]', img: 'https://images.pokemontcg.io/base1/2_hier.png' },
-      { id: 'base1-15', name: 'Venusaur', set_id: 'base1', set_name: 'Base Set', number: '15', rarity: 'Rare Holo', supertype: 'Pokémon', types: '["Grass"]', img: 'https://images.pokemontcg.io/base1/15_hier.png' },
-      { id: 'base1-10', name: 'Mewtwo', set_id: 'base1', set_name: 'Base Set', number: '10', rarity: 'Rare Holo', supertype: 'Pokémon', types: '["Psychic"]', img: 'https://images.pokemontcg.io/base1/10_hier.png' },
-      { id: 'fossil-20', name: 'Gengar', set_id: 'fossil', set_name: 'Fossil', number: '20', rarity: 'Rare', supertype: 'Pokémon', types: '["Psychic"]', img: 'https://images.pokemontcg.io/fossil/20_hier.png' },
-      { id: 'jungle-51', name: 'Eevee', set_id: 'jungle', set_name: 'Jungle', number: '51', rarity: 'Common', supertype: 'Pokémon', types: '["Colorless"]', img: 'https://images.pokemontcg.io/jungle/51_hier.png' },
-      { id: 'neo1-9', name: 'Lugia', set_id: 'neo1', set_name: 'Neo Genesis', number: '9', rarity: 'Rare Holo', supertype: 'Pokémon', types: '["Colorless"]', img: 'https://images.pokemontcg.io/neo1/9_hier.png' },
-      { id: 'ex3-102', name: 'Rayquaza', set_id: 'ex3', set_name: 'EX Deoxys', number: '102', rarity: 'Rare Holo Star', supertype: 'Pokémon', types: '["Colorless"]', img: 'https://images.pokemontcg.io/ex3/102_hier.png' },
-      { id: 'base1-6', name: 'Gyarados', set_id: 'base1', set_name: 'Base Set', number: '6', rarity: 'Rare Holo', supertype: 'Pokémon', types: '["Water"]', img: 'https://images.pokemontcg.io/base1/6_hier.png' }
-    ];
-
-    // Seed mock cards into card_cache
-    for (const card of MOCK_POOL) {
-      const priceNormal = parseFloat((1.5 + Math.random() * 5).toFixed(2));
-      const priceHolo = parseFloat((10 + Math.random() * 60).toFixed(2));
-      const priceRev = parseFloat((5 + Math.random() * 30).toFixed(2));
-      
-      await db.run(`
-        INSERT OR REPLACE INTO card_cache (id, name, supertype, subtypes, types, rarity, set_id, set_name, number, image_url, price_trend, price_normal, price_holofoil, price_reverse_holofoil)
-        VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-      `, [card.id, card.name, card.supertype, '["Basic"]', card.types, card.rarity, card.set_id, card.set_name, card.number, card.img, priceHolo, priceNormal, priceHolo, priceRev]);
-    }
-
-    // Insert random collection entries distributed across binder & box
-    const prints = ['Normal', 'Holofoil', 'Reverse Holofoil'];
-    const conditions = ['Near Mint', 'Lightly Played', 'Moderately Played', 'Heavily Played'];
-    
-    let addedCount = 0;
-    
-    // Seed binder entries (use pages 1 to 3, slots 1 to 9)
-    for (let p = 1; p <= 3; p++) {
-      for (let s = 1; s <= 9; s++) {
-        if (Math.random() > 0.25) {
-          const card = MOCK_POOL[Math.floor(Math.random() * MOCK_POOL.length)];
-          const print = prints[Math.floor(Math.random() * prints.length)];
-          const condition = conditions[Math.floor(Math.random() * conditions.length)];
-          const qty = Math.floor(Math.random() * 3) + 1; // 1-3 copies
-          const purchasePrice = parseFloat((Math.random() * 10).toFixed(2));
-          const pos = ((p - 1) * 9 + (s - 1)) * 1000;
-          
-          await db.run(`
-            INSERT INTO collection (card_id, quantity, condition, printing, language, purchase_price, location_id, sub_location_1, sub_location_2, position, user_id)
-            VALUES (?, ?, ?, ?, 'English', ?, ?, ?, ?, ?, ?)
-          `, [card.id, qty, condition, print, purchasePrice, binder.id, `Page ${p}`, `Slot ${s}`, pos, req.user.id]);
-          addedCount += qty;
-        }
-      }
-    }
-
-    // Seed box entries (use row 1-2, slot 1-15)
-    for (let r = 1; r <= 2; r++) {
-      for (let s = 1; s <= 15; s++) {
-        if (Math.random() > 0.4) {
-          const card = MOCK_POOL[Math.floor(Math.random() * MOCK_POOL.length)];
-          const print = prints[Math.floor(Math.random() * prints.length)];
-          const condition = conditions[Math.floor(Math.random() * conditions.length)];
-          const qty = Math.floor(Math.random() * 4) + 1; // 1-4 copies
-          const purchasePrice = parseFloat((Math.random() * 5).toFixed(2));
-          const pos = ((r - 1) * 40 + (s - 1)) * 1000;
-          
-          await db.run(`
-            INSERT INTO collection (card_id, quantity, condition, printing, language, purchase_price, location_id, sub_location_1, sub_location_2, position, user_id)
-            VALUES (?, ?, ?, ?, 'English', ?, ?, ?, ?, ?, ?)
-          `, [card.id, qty, condition, print, purchasePrice, box.id, `Row ${r}`, `Slot ${s}`, pos, req.user.id]);
-          addedCount += qty;
-        }
-      }
-    }
-
-    res.json({ message: `Successfully seeded test collection with ${addedCount} cards for admin user!` });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to seed test cards' });
-  }
-});
-
-// --- CARD MANAGEMENT ENDPOINTS ---
-
-// 1. Search Pokémon TCG cards (proxies to Pokemon TCG API and database cache)
-app.get('/api/search', authenticateToken, async (req, res) => {
-  const { name, number, set } = req.query;
-  try {
-    const results = await tcgApi.searchCards(name, number, set, req.user.tcg_api_key);
-    res.json(results);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Search failed' });
-  }
-});
-
-// 2. Get User's Collection
-app.get('/api/collection', authenticateToken, async (req, res) => {
-  try {
-    const listType = req.query.list_type || 'collection';
-    const isTrade = req.query.is_trade;
-
-    let filterSql = `WHERE c.user_id = ? AND c.list_type = ?`;
-    let filterParams = [req.user.id, listType];
-
-    if (isTrade !== undefined) {
-      filterSql += ` AND c.is_trade = ?`;
-      filterParams.push(isTrade === 'true' || isTrade === '1' ? 1 : 0);
-    }
-
-    const query = `
-      SELECT 
-        c.id as entry_id,
-        c.card_id,
-        c.quantity,
-        c.condition,
-        c.printing,
-        c.language,
-        c.purchase_price,
-        c.sub_location_1,
-        c.sub_location_2,
-        c.position,
-        c.added_at,
-        c.is_trade,
-        c.list_type,
-        cc.name,
-        cc.supertype,
-        cc.subtypes,
-        cc.types,
-        cc.rarity,
-        cc.set_id,
-        cc.set_name,
-        cc.number,
-        cc.image_url,
-        cc.price_trend,
-        cc.price_normal,
-        cc.price_holofoil,
-        cc.price_reverse_holofoil,
-        l.id as location_id,
-        l.name as location_name,
-        l.type as location_type
-      FROM collection c
-      JOIN card_cache cc ON c.card_id = cc.id
-      LEFT JOIN locations l ON c.location_id = l.id
-      ${filterSql}
-      ORDER BY c.added_at DESC
-    `;
-    const rows = await db.all(query, filterParams);
-    
-    // Parse JSON fields
-    const formatted = rows.map(row => ({
-      ...row,
-      price_trend: resolveCardPrice(row),
-      subtypes: JSON.parse(row.subtypes || '[]'),
-      types: JSON.parse(row.types || '[]'),
-    }));
-    
-    res.json(formatted);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to retrieve collection' });
-  }
-});
-
-// 3. Add Card to Collection
-app.post('/api/collection', authenticateToken, async (req, res) => {
-  const {
-    card_id,
-    quantity = 1,
-    condition = 'Near Mint',
-    printing = 'Normal',
-    language = 'English',
-    purchase_price = 0,
-    location_id = null,
-    sub_location_1 = '',
-    sub_location_2 = '',
-    list_type = 'collection',
-    is_trade = 0,
-    position
-  } = req.body;
-
-  if (!card_id) {
-    return res.status(400).json({ error: 'card_id is required' });
-  }
-
-  try {
-    // Ensure location_id belongs to the user if provided
-    if (location_id) {
-      const loc = await db.get(`SELECT id FROM locations WHERE id = ? AND user_id = ?`, [location_id, req.user.id]);
-      if (!loc) {
-        return res.status(400).json({ error: 'Invalid location ID' });
-      }
-    }
-
-    // Ensure card is in the local metadata cache
-    let card = await db.get(`SELECT id FROM card_cache WHERE id = ?`, [card_id]);
-    if (!card) {
-      console.log(`Card ${card_id} not found in cache. Fetching from API first...`);
-      const apiCard = await tcgApi.getCardById(card_id, req.user.tcg_api_key);
-      if (!apiCard) {
-        return res.status(404).json({ error: `Card ID ${card_id} not found on Pokémon TCG API.` });
-      }
-    }
-
-    // Check if an identical card entry already exists in this location to stack them
-    const existing = await db.get(`
-      SELECT id, quantity FROM collection 
-      WHERE card_id = ? AND condition = ? AND printing = ? AND language = ? 
-        AND location_id IS ? AND sub_location_1 IS ? AND sub_location_2 IS ?
-        AND user_id = ? AND list_type = ? AND is_trade = ?
-    `, [
-      card_id,
-      condition,
-      printing,
-      language,
-      location_id,
-      sub_location_1 || null,
-      sub_location_2 || null,
-      req.user.id,
-      list_type,
-      is_trade ? 1 : 0
-    ]);
-
-    let lastInsertedId;
-    if (existing) {
-      const newQuantity = existing.quantity + parseInt(quantity, 10);
-      await db.run(`
-        UPDATE collection SET quantity = ? WHERE id = ?
-      `, [newQuantity, existing.id]);
-      lastInsertedId = existing.id;
-    } else {
-      let finalPosition = position;
-      if (finalPosition === undefined) {
-        if (location_id) {
-          const cardMetadata = await db.get(`SELECT name, set_name, number, types, price_trend, supertype, rarity FROM card_cache WHERE id = ?`, [card_id]);
-          if (cardMetadata) {
-            cardMetadata.printing = printing;
-            try { cardMetadata.types = JSON.parse(cardMetadata.types || '[]'); } catch { cardMetadata.types = []; }
-            finalPosition = await getSortedPositionForCard(db, location_id, req.user.id, cardMetadata);
-          } else {
-            finalPosition = 1000;
-          }
-        } else {
-          finalPosition = 0;
-        }
-      }
-
-      const result = await db.run(`
-        INSERT INTO collection 
-        (card_id, quantity, condition, printing, language, purchase_price, location_id, sub_location_1, sub_location_2, user_id, list_type, is_trade, position)
-        VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-      `, [
-        card_id,
-        quantity,
-        condition,
-        printing,
-        language,
-        purchase_price || 0,
-        location_id,
-        sub_location_1 || null,
-        sub_location_2 || null,
-        req.user.id,
-        list_type,
-        is_trade ? 1 : 0,
-        finalPosition
-      ]);
-      lastInsertedId = result.lastID;
-
-      if (location_id) {
-        const neighbors = await db.all(`
-          SELECT position FROM collection 
-          WHERE location_id = ? AND user_id = ? AND id != ?
-          ORDER BY ABS(position - ?) ASC LIMIT 2
-        `, [location_id, req.user.id, lastInsertedId, finalPosition]);
-        let needsRebalance = false;
-        for (const n of neighbors) {
-          if (Math.abs(n.position - finalPosition) < 0.001) {
-            needsRebalance = true;
-            break;
-          }
-        }
-        if (needsRebalance) {
-          await rebalanceLocationPositions(db, location_id, req.user.id);
-        }
-      }
-    }
-
-    // Record initial price history trend
-    const cacheCard = await db.get(`SELECT price_trend FROM card_cache WHERE id = ?`, [card_id]);
-    if (cacheCard && cacheCard.price_trend > 0) {
-      await db.run(`INSERT OR IGNORE INTO price_history (card_id, price) VALUES (?, ?)`, [card_id, cacheCard.price_trend]);
-    }
-
-    res.status(201).json({ message: 'Card added to collection', id: lastInsertedId });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to add card' });
-  }
-});
-
-// 4. Update Card in Collection
-app.put('/api/collection/:id', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  const {
-    quantity,
-    condition,
-    printing,
-    language,
-    purchase_price,
-    location_id,
-    sub_location_1,
-    sub_location_2,
-    list_type,
-    is_trade,
-    position
-  } = req.body;
-
-  try {
-    // Ensure entry exists and belongs to the user
-    const entry = await db.get(`SELECT * FROM collection WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (!entry) {
-      return res.status(404).json({ error: 'Collection entry not found' });
-    }
-
-    // Ensure location_id belongs to the user if updated
-    if (location_id) {
-      const loc = await db.get(`SELECT id FROM locations WHERE id = ? AND user_id = ?`, [location_id, req.user.id]);
-      if (!loc) {
-        return res.status(400).json({ error: 'Invalid location ID' });
-      }
-    }
-
-    const finalLocId = location_id !== undefined ? location_id : entry.location_id;
-    const finalSub1 = sub_location_1 !== undefined ? sub_location_1 : entry.sub_location_1;
-    const finalSub2 = sub_location_2 !== undefined ? sub_location_2 : entry.sub_location_2;
-
-    // Compute what the final values would be after this update
-    const finalQuantity = quantity !== undefined ? parseInt(quantity, 10) : entry.quantity;
-    const finalCondition = condition !== undefined ? condition : entry.condition;
-    const finalPrinting = printing !== undefined ? printing : entry.printing;
-    const finalLanguage = language !== undefined ? language : entry.language;
-    const finalListType = list_type !== undefined ? list_type : entry.list_type;
-    const finalIsTrade = is_trade !== undefined ? (is_trade ? 1 : 0) : entry.is_trade;
-
-    // Check if there is an identical card entry already in that new location (except this row itself)
-    const existing = await db.get(`
-      SELECT id, quantity FROM collection 
-      WHERE card_id = ? AND condition = ? AND printing = ? AND language = ? 
-        AND location_id IS ? AND sub_location_1 IS ? AND sub_location_2 IS ?
-        AND user_id = ? AND list_type = ? AND is_trade = ? AND id != ?
-    `, [
-      entry.card_id,
-      finalCondition,
-      finalPrinting,
-      finalLanguage,
-      finalLocId,
-      finalSub1 || null,
-      finalSub2 || null,
-      req.user.id,
-      finalListType,
-      finalIsTrade,
-      entry.id
-    ]);
-
-    if (existing) {
-      const newQuantity = existing.quantity + finalQuantity;
-      await db.run(`UPDATE collection SET quantity = ? WHERE id = ?`, [newQuantity, existing.id]);
-      await db.run(`DELETE FROM collection WHERE id = ? AND user_id = ?`, [entry.id, req.user.id]);
-      return res.json({ message: 'Collection entry merged/stacked successfully', id: existing.id });
-    }
-
-    let finalPosition = position;
-    if (finalPosition === undefined && location_id !== undefined && location_id !== entry.location_id) {
-      if (location_id) {
-        const cardMetadata = await db.get(`SELECT name, set_name, number, types, price_trend, supertype, rarity FROM card_cache WHERE id = ?`, [entry.card_id]);
-        if (cardMetadata) {
-          cardMetadata.printing = printing !== undefined ? printing : entry.printing;
-          try { cardMetadata.types = JSON.parse(cardMetadata.types || '[]'); } catch { cardMetadata.types = []; }
-          finalPosition = await getSortedPositionForCard(db, location_id, req.user.id, cardMetadata);
-        } else {
-          finalPosition = 1000;
-        }
-      } else {
-        finalPosition = 0;
-      }
-    }
-
-    // Build dynamic UPDATE query based on passed values
-    const fields = [];
-    const params = [];
-
-    if (quantity !== undefined) { fields.push('quantity = ?'); params.push(quantity); }
-    if (condition !== undefined) { fields.push('condition = ?'); params.push(condition); }
-    if (printing !== undefined) { fields.push('printing = ?'); params.push(printing); }
-    if (language !== undefined) { fields.push('language = ?'); params.push(language); }
-    if (purchase_price !== undefined) { fields.push('purchase_price = ?'); params.push(purchase_price); }
-    if (location_id !== undefined) { fields.push('location_id = ?'); params.push(location_id); }
-    if (sub_location_1 !== undefined) { fields.push('sub_location_1 = ?'); params.push(sub_location_1); }
-    if (sub_location_2 !== undefined) { fields.push('sub_location_2 = ?'); params.push(sub_location_2); }
-    if (list_type !== undefined) { fields.push('list_type = ?'); params.push(list_type); }
-    if (is_trade !== undefined) { fields.push('is_trade = ?'); params.push(is_trade ? 1 : 0); }
-    if (finalPosition !== undefined) { fields.push('position = ?'); params.push(finalPosition); }
-
-    if (fields.length === 0) {
-      return res.status(400).json({ error: 'No fields provided for update' });
-    }
-
-    params.push(id);
-    params.push(req.user.id);
-    await db.run(`UPDATE collection SET ${fields.join(', ')} WHERE id = ? AND user_id = ?`, params);
-
-    if (finalLocId && finalPosition !== undefined) {
-      const neighbors = await db.all(`
-        SELECT position FROM collection 
-        WHERE location_id = ? AND user_id = ? AND id != ?
-        ORDER BY ABS(position - ?) ASC LIMIT 2
-      `, [finalLocId, req.user.id, id, finalPosition]);
-      let needsRebalance = false;
-      for (const n of neighbors) {
-        if (Math.abs(n.position - finalPosition) < 0.001) {
-          needsRebalance = true;
-          break;
-        }
-      }
-      if (needsRebalance) {
-        await rebalanceLocationPositions(db, finalLocId, req.user.id);
-      }
-    }
-    
-    res.json({ message: 'Collection entry updated successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to update entry' });
-  }
-});
-
-// 5. Delete Card from Collection
-app.delete('/api/collection/:id', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  try {
-    const result = await db.run(`DELETE FROM collection WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (result.changes === 0) {
-      return res.status(404).json({ error: 'Collection entry not found' });
-    }
-    res.json({ message: 'Card removed from collection' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to remove card' });
-  }
-});
-
-// 6. Manage Locations (Physical Storage)
-app.get('/api/locations', authenticateToken, async (req, res) => {
-  try {
-    const locations = await db.all(`
-      SELECT l.*, COUNT(c.id) as card_count, SUM(c.quantity) as total_cards 
-      FROM locations l
-      LEFT JOIN collection c ON l.id = c.location_id AND c.user_id = l.user_id
-      WHERE l.user_id = ?
-      GROUP BY l.id
-    `, [req.user.id]);
-    res.json(locations);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to retrieve locations' });
-  }
-});
-
-app.post('/api/locations', authenticateToken, async (req, res) => {
-  const { 
-    name, 
-    type, 
-    description = '', 
-    sort_order = 'name-asc',
-    max_pages = 30,
-    page_style = '3x3',
-    max_rows = 3,
-    max_capacity = 1000,
-    foil_sorting = 'normals_first'
-  } = req.body;
-
-  if (!name || !type) {
-    return res.status(400).json({ error: 'name and type are required' });
-  }
-  try {
-    const existing = await db.get(`SELECT id FROM locations WHERE name = ? AND user_id = ?`, [name, req.user.id]);
-    if (existing) {
-      return res.status(400).json({ error: 'A location with this name already exists' });
-    }
-
-    const result = await db.run(`
-      INSERT INTO locations (name, type, description, sort_order, max_pages, page_style, max_rows, max_capacity, foil_sorting, user_id)
-      VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-    `, [
-      name, 
-      type, 
-      description, 
-      sort_order, 
-      parseInt(max_pages, 10) || 30, 
-      page_style, 
-      parseInt(max_rows, 10) || 3, 
-      parseInt(max_capacity, 10) || 1000, 
-      foil_sorting || 'normals_first',
-      req.user.id
-    ]);
-    res.status(201).json({ message: 'Location created', id: result.lastID });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to create location' });
-  }
-});
-
-app.put('/api/locations/:id', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  const { name, type, description, sort_order, max_pages, page_style, max_rows, max_capacity, foil_sorting } = req.body;
-  try {
-    const loc = await db.get(`SELECT id FROM locations WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (!loc) {
-      return res.status(404).json({ error: 'Location not found' });
-    }
-
-    if (name) {
-      const dup = await db.get(`SELECT id FROM locations WHERE name = ? AND user_id = ? AND id != ?`, [name, req.user.id, id]);
-      if (dup) {
-        return res.status(400).json({ error: 'A location with this name already exists' });
-      }
-    }
-
-    await db.run(`
-      UPDATE locations 
-      SET 
-        name = COALESCE(?, name), 
-        type = COALESCE(?, type), 
-        description = COALESCE(?, description),
-        sort_order = COALESCE(?, sort_order),
-        max_pages = COALESCE(?, max_pages),
-        page_style = COALESCE(?, page_style),
-        max_rows = COALESCE(?, max_rows),
-        max_capacity = COALESCE(?, max_capacity),
-        foil_sorting = COALESCE(?, foil_sorting)
-      WHERE id = ? AND user_id = ?
-    `, [
-      name, 
-      type, 
-      description, 
-      sort_order, 
-      max_pages !== undefined ? parseInt(max_pages, 10) : null,
-      page_style,
-      max_rows !== undefined ? parseInt(max_rows, 10) : null,
-      max_capacity !== undefined ? parseInt(max_capacity, 10) : null,
-      foil_sorting,
-      id, 
-      req.user.id
-    ]);
-    res.json({ message: 'Location updated' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to update location' });
-  }
-});
-
-app.delete('/api/locations/:id', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  try {
-    const loc = await db.get(`SELECT id FROM locations WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (!loc) {
-      return res.status(404).json({ error: 'Location not found' });
-    }
-
-    // Disassociate cards from this location instead of blocking delete (scoped to user)
-    await db.run(`UPDATE collection SET location_id = NULL, sub_location_1 = NULL, sub_location_2 = NULL WHERE location_id = ? AND user_id = ?`, [id, req.user.id]);
-    
-    await db.run(`DELETE FROM locations WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    res.json({ message: 'Location deleted successfully (any stored cards moved to Unsorted)' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to delete location' });
-  }
-});
-
-// Helper for deterministic price simulation over history
-const getSimulatedPriceAt = (cardId, currentPrice, t, now) => {
-  const timeDiff = now - t;
-  const yearsAgo = timeDiff / (365.25 * 24 * 60 * 60 * 1000);
-  
-  // Deterministic seed from card ID
-  let seed = 0;
-  for (let i = 0; i < cardId.length; i++) {
-    seed += cardId.charCodeAt(i);
-  }
-  
-  // Simulate a slow appreciation/depreciation over the years with some sin wave noise
-  const trend = 0.05 * Math.sin(seed * 0.1) * yearsAgo; // up or down trend
-  const noise = 0.03 * Math.sin(seed + yearsAgo * 12); // monthly waves
-  const factor = Math.max(0.1, 1.0 - (yearsAgo * 0.06) + trend + noise);
-  
-  return parseFloat((currentPrice * factor).toFixed(2));
-};
-
-const isVintageSet = (setId) => {
-  const id = (setId || '').toLowerCase();
-  return id.startsWith('base') || id.startsWith('gym') || id.startsWith('neo') || 
-         id.startsWith('lc') || id.startsWith('ecard') || id.startsWith('ex') || 
-         id.startsWith('pop') || id.startsWith('promo1') || id.startsWith('si') ||
-         id.startsWith('xy12') || id.startsWith('cel25');
-};
-
-// 7. Get Collection Statistics & Analytics
-app.get('/api/stats', authenticateToken, async (req, res) => {
-  try {
-    // Retrieve all collection items to compute statistics
-    const query = `
-      SELECT 
-        c.quantity, c.purchase_price, c.added_at, c.printing, c.condition, c.card_id,
-        cc.types, cc.rarity, cc.set_name, cc.set_id, cc.price_trend, cc.price_normal, cc.price_holofoil, cc.price_reverse_holofoil,
-        l.name as location_name
-      FROM collection c
-      JOIN card_cache cc ON c.card_id = cc.id
-      LEFT JOIN locations l ON c.location_id = l.id
-      WHERE c.user_id = ?
-    `;
-    const rows = await db.all(query, [req.user.id]);
-
-    let totalCards = 0;
-    let uniqueCards = rows.length;
-    let totalValue = 0;
-    let nearMintCount = 0;
-    let vintageCount = 0;
-
-    const now = Date.now();
-    const oneDayMs = 24 * 60 * 60 * 1000;
-    const sevenDaysMs = 7 * oneDayMs;
-    const thirtyDaysMs = 30 * oneDayMs;
-    const oneYearMs = 365 * oneDayMs;
-    const fiveYearsMs = 5 * oneYearMs;
-
-    let value7dAgo = 0;
-    let value30dAgo = 0;
-    let value1yAgo = 0;
-    let value5yAgo = 0;
-
-    const typeCounts = {};
-    const rarityCounts = {};
-    const setCounts = {};
-    const locationCounts = {};
-
-    rows.forEach(row => {
-      const qty = row.quantity || 1;
-      const price = resolveCardPrice(row);
-      const addedTime = row.added_at ? new Date(row.added_at).getTime() : now;
-
-      totalCards += qty;
-      totalValue += qty * price;
-
-      if (row.condition === 'Near Mint') {
-        nearMintCount += qty;
-      }
-
-      if (isVintageSet(row.set_id)) {
-        vintageCount += qty;
-      }
-
-      // Check if the card was in the collection at those times
-      if (addedTime <= now - sevenDaysMs) {
-        value7dAgo += qty * getSimulatedPriceAt(row.card_id, price, now - sevenDaysMs, now);
-      }
-      if (addedTime <= now - thirtyDaysMs) {
-        value30dAgo += qty * getSimulatedPriceAt(row.card_id, price, now - thirtyDaysMs, now);
-      }
-      if (addedTime <= now - oneYearMs) {
-        value1yAgo += qty * getSimulatedPriceAt(row.card_id, price, now - oneYearMs, now);
-      }
-      if (addedTime <= now - fiveYearsMs) {
-        value5yAgo += qty * getSimulatedPriceAt(row.card_id, price, now - fiveYearsMs, now);
-      }
-
-      // Parse types
-      const types = JSON.parse(row.types || '[]');
-      types.forEach(t => {
-        typeCounts[t] = (typeCounts[t] || 0) + qty;
-      });
-      if (types.length === 0) {
-        typeCounts['Colorless'] = (typeCounts['Colorless'] || 0) + qty;
-      }
-
-      // Rarity
-      const rarity = row.rarity || 'Unknown';
-      rarityCounts[rarity] = (rarityCounts[rarity] || 0) + qty;
-
-      // Set
-      const set = row.set_name || 'Other';
-      if (!setCounts[row.set_id]) {
-        setCounts[row.set_id] = { name: set, count: 0, value: 0 };
-      }
-      setCounts[row.set_id].count += qty;
-      setCounts[row.set_id].value += qty * price;
-
-      // Location
-      const loc = row.location_name || 'Unassigned';
-      locationCounts[loc] = (locationCounts[loc] || 0) + qty;
-    });
-
-    // Get top most valuable cards (scoped to user)
-    const topValuableQuery = `
-      SELECT 
-        c.quantity, c.condition, c.printing, c.language, c.purchase_price,
-        cc.id as card_id, cc.name, cc.rarity, cc.set_name, cc.image_url, cc.price_trend,
-        cc.price_normal, cc.price_holofoil, cc.price_reverse_holofoil
-      FROM collection c
-      JOIN card_cache cc ON c.card_id = cc.id
-      WHERE c.user_id = ?
-      ORDER BY CASE 
-        WHEN c.printing = 'Holofoil' AND cc.price_holofoil IS NOT NULL AND cc.price_holofoil > 0 THEN cc.price_holofoil
-        WHEN c.printing = 'Reverse Holofoil' AND cc.price_reverse_holofoil IS NOT NULL AND cc.price_reverse_holofoil > 0 THEN cc.price_reverse_holofoil
-        WHEN c.printing = 'Normal' AND cc.price_normal IS NOT NULL AND cc.price_normal > 0 THEN cc.price_normal
-        ELSE cc.price_trend
-      END DESC
-      LIMIT 6
-    `;
-    const topValuableRows = await db.all(topValuableQuery, [req.user.id]);
-    const topValuable = topValuableRows.map(row => ({
-      ...row,
-      price_trend: resolveCardPrice(row)
-    }));
-
-    // Compute progress for top 4 sets in database (estimate set total)
-    const setSizes = {
-      'base1': 102,  // Base Set
-      'base2': 64,   // Jungle
-      'base3': 62,   // Fossil
-      'base4': 130,  // Base Set 2
-      'neo1': 111,   // Neo Genesis
-      'cel25': 25,   // Celebrations
-      'swsh1': 202,  // Sword & Shield Base
-      'swsh11': 196, // Lost Origin
-      'swsh12': 98,  // Silver Tempest
-      'sv1': 198,    // Scarlet & Violet Base
-      'sv2': 193,    // Paldea Evolved
-      'sv3': 197,    // Obsidian Flames
-      'sv3pt5': 165, // 151
-    };
-
-    const setProgress = [];
-    for (const setId in setCounts) {
-      const userUniqueInSet = await db.get(`
-        SELECT COUNT(DISTINCT card_id) as count 
-        FROM collection c
-        JOIN card_cache cc ON c.card_id = cc.id
-        WHERE cc.set_id = ? AND c.user_id = ?
-      `, [setId, req.user.id]);
-
-      const size = setSizes[setId] || 150; // default estimate if set not in database
-      const count = userUniqueInSet.count;
-      setProgress.push({
-        setId,
-        setName: setCounts[setId].name,
-        ownedUnique: count,
-        totalCards: size,
-        percent: Math.min(Math.round((count / size) * 100), 100)
-      });
-    }
-
-    // Sort set progress by completion percentage descending
-    setProgress.sort((a, b) => b.percent - a.percent);
-
-    const mintRate = totalCards > 0 ? parseFloat(((nearMintCount / totalCards) * 100).toFixed(1)) : 0.0;
-    const vintageRatio = totalCards > 0 ? parseFloat(((vintageCount / totalCards) * 100).toFixed(1)) : 0.0;
-
-    res.json({
-      summary: {
-        totalCards,
-        uniqueCards,
-        totalValue: parseFloat(totalValue.toFixed(2)),
-        mintRate,
-        vintageRatio,
-        change7d: {
-          abs: parseFloat((totalValue - value7dAgo).toFixed(2)),
-          pct: value7dAgo > 0 ? parseFloat((((totalValue - value7dAgo) / value7dAgo) * 100).toFixed(1)) : 100.0
-        },
-        change30d: {
-          abs: parseFloat((totalValue - value30dAgo).toFixed(2)),
-          pct: value30dAgo > 0 ? parseFloat((((totalValue - value30dAgo) / value30dAgo) * 100).toFixed(1)) : 100.0
-        },
-        change1y: {
-          abs: parseFloat((totalValue - value1yAgo).toFixed(2)),
-          pct: value1yAgo > 0 ? parseFloat((((totalValue - value1yAgo) / value1yAgo) * 100).toFixed(1)) : 100.0
-        },
-        change5y: {
-          abs: parseFloat((totalValue - value5yAgo).toFixed(2)),
-          pct: value5yAgo > 0 ? parseFloat((((totalValue - value5yAgo) / value5yAgo) * 100).toFixed(1)) : 100.0
-        }
-      },
-      types: Object.keys(typeCounts).map(name => ({ name, value: typeCounts[name] })),
-      rarities: Object.keys(rarityCounts).map(name => ({ name, value: rarityCounts[name] })),
-      sets: Object.keys(setCounts).map(id => ({ 
-        id, 
-        name: setCounts[id].name, 
-        count: setCounts[id].count, 
-        value: parseFloat(setCounts[id].value.toFixed(2)) 
-      })).sort((a,b) => b.value - a.value).slice(0, 8),
-      locations: Object.keys(locationCounts).map(name => ({ name, value: locationCounts[name] })),
-      topValuable,
-      setProgress: setProgress.slice(0, 4)
-    });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to compute statistics' });
-  }
-});
-
-// 7b. Get Collection Net Worth Timeline History
-app.get('/api/stats/history', authenticateToken, async (req, res) => {
-  try {
-    const { period = '30d' } = req.query;
-    
-    // Retrieve all collection items to compute history
-    const query = `
-      SELECT c.quantity, c.added_at, c.printing, cc.id as card_id, cc.price_trend, cc.price_normal, cc.price_holofoil, cc.price_reverse_holofoil
-      FROM collection c
-      JOIN card_cache cc ON c.card_id = cc.id
-      WHERE c.user_id = ?
-    `;
-    const items = await db.all(query, [req.user.id]);
-
-    const now = Date.now();
-    let points = [];
-    let step = 0;
-    let count = 0;
-    let formatLabel = (d) => d.toLocaleDateString();
-
-    if (period === '7d') {
-      count = 7;
-      step = 24 * 60 * 60 * 1000;
-      formatLabel = (d) => d.toLocaleDateString(undefined, { weekday: 'short' });
-    } else if (period === '30d') {
-      count = 30;
-      step = 24 * 60 * 60 * 1000;
-      formatLabel = (d) => d.toLocaleDateString(undefined, { month: 'short', day: 'numeric' });
-    } else if (period === '1y') {
-      count = 12;
-      step = 30 * 24 * 60 * 60 * 1000;
-      formatLabel = (d) => d.toLocaleDateString(undefined, { month: 'short', year: '2-digit' });
-    } else if (period === '5y') {
-      count = 20;
-      step = 91 * 24 * 60 * 60 * 1000;
-      formatLabel = (d) => d.toLocaleDateString(undefined, { month: 'short', year: 'numeric' });
-    } else {
-      count = 30;
-      step = 24 * 60 * 60 * 1000;
-      formatLabel = (d) => d.toLocaleDateString(undefined, { month: 'short', day: 'numeric' });
-    }
-
-    const historyData = [];
-    for (let i = count - 1; i >= 0; i--) {
-      const targetTime = now - (i * step);
-      const targetDate = new Date(targetTime);
-      
-      let totalValue = 0;
-      items.forEach(item => {
-        const addedTime = new Date(item.added_at).getTime();
-        if (addedTime <= targetTime) {
-          const currentPrice = resolveCardPrice(item);
-          const simulatedPrice = getSimulatedPriceAt(item.card_id, currentPrice, targetTime, now);
-          totalValue += item.quantity * simulatedPrice;
-        }
-      });
-
-      historyData.push({
-        date: formatLabel(targetDate),
-        value: parseFloat(totalValue.toFixed(2))
-      });
-    }
-
-    res.json(historyData);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to compute timeline history' });
-  }
-});
-
-// 8. Export Database
-app.get('/api/export', authenticateToken, async (req, res) => {
-  const { format = 'csv' } = req.query;
-  try {
-    const query = `
-      SELECT 
-        c.quantity,
-        c.condition,
-        c.printing,
-        c.language,
-        c.purchase_price,
-        c.sub_location_1,
-        c.sub_location_2,
-        c.added_at,
-        cc.id as card_id,
-        cc.name as card_name,
-        cc.supertype,
-        cc.types,
-        cc.rarity,
-        cc.set_id,
-        cc.set_name,
-        cc.number as card_number,
-        cc.image_url,
-        cc.price_trend,
-        cc.price_normal,
-        cc.price_holofoil,
-        cc.price_reverse_holofoil,
-        l.name as location_name
-      FROM collection c
-      JOIN card_cache cc ON c.card_id = cc.id
-      LEFT JOIN locations l ON c.location_id = l.id
-      WHERE c.user_id = ?
-    `;
-    const dbRows = await db.all(query, [req.user.id]);
-    const rows = dbRows.map(row => {
-      const resolvedPrice = resolveCardPrice(row);
-      return {
-        ...row,
-        market_price: resolvedPrice
-      };
-    });
-
-    if (format.toLowerCase() === 'json') {
-      res.setHeader('Content-Type', 'application/json');
-      res.setHeader('Content-Disposition', 'attachment; filename=pokedexrr_collection.json');
-      return res.json(rows);
-    }
-
-    // Default to CSV
-    res.setHeader('Content-Type', 'text/csv');
-    res.setHeader('Content-Disposition', 'attachment; filename=pokedexrr_collection.csv');
-
-    // Headers
-    const headers = [
-      'Card ID', 'Name', 'Set Name', 'Set ID', 'Card Number', 'Rarity',
-      'Quantity', 'Condition', 'Printing', 'Language', 'Purchase Price',
-      'Market Price', 'Location Container', 'Sub-Location Page/Row', 'Sub-Location Slot/Section', 'Added At'
-    ];
-
-    // Neutralize leading =, +, -, @ so spreadsheet apps don't interpret free-text
-    // fields (card/location names) as formulas when the export is opened.
-    const csvCell = (value) => {
-      const str = String(value ?? '');
-      return /^[=+\-@]/.test(str) ? `'${str}` : str;
-    };
-
-    let csvContent = headers.join(',') + '\n';
-
-    rows.forEach(r => {
-      const line = [
-        r.card_id,
-        `"${csvCell(r.card_name).replace(/"/g, '""')}"`,
-        `"${csvCell(r.set_name).replace(/"/g, '""')}"`,
-        r.set_id,
-        r.card_number,
-        r.rarity,
-        r.quantity,
-        r.condition,
-        r.printing,
-        r.language,
-        r.purchase_price || 0,
-        r.market_price || 0,
-        r.location_name ? `"${csvCell(r.location_name).replace(/"/g, '""')}"` : 'Unassigned',
-        r.sub_location_1 ? `"${csvCell(r.sub_location_1).replace(/"/g, '""')}"` : '',
-        r.sub_location_2 ? `"${csvCell(r.sub_location_2).replace(/"/g, '""')}"` : '',
-        r.added_at
-      ];
-      csvContent += line.join(',') + '\n';
-    });
-
-    res.send(csvContent);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Export failed' });
-  }
-});
-
-// 8b. Import Database
-app.post('/api/import', authenticateToken, async (req, res) => {
-  try {
-    const { format, data } = req.body;
-    if (!data) {
-      return res.status(400).json({ error: 'No data provided' });
-    }
-
-    let cards = [];
-
-    if (format === 'json') {
-      cards = typeof data === 'string' ? JSON.parse(data) : data;
-    } else if (format === 'csv') {
-      const lines = data.split('\n').map(l => l.trim()).filter(Boolean);
-      if (lines.length <= 1) {
-        return res.status(400).json({ error: 'CSV file is empty' });
-      }
-
-      const headers = lines[0].split(',').map(h => h.trim().replace(/"/g, ''));
-      
-      const parseCSVLine = (line) => {
-        const result = [];
-        let current = '';
-        let inQuotes = false;
-        for (let i = 0; i < line.length; i++) {
-          const char = line[i];
-          if (char === '"') {
-            inQuotes = !inQuotes;
-          } else if (char === ',' && !inQuotes) {
-            result.push(current.trim());
-            current = '';
-          } else {
-            current += char;
-          }
-        }
-        result.push(current.trim());
-        return result;
-      };
-
-      for (let i = 1; i < lines.length; i++) {
-        const values = parseCSVLine(lines[i]);
-        if (values.length < headers.length) continue;
-
-        const cardObj = {};
-        headers.forEach((header, idx) => {
-          cardObj[header] = values[idx];
-        });
-        
-        cards.push({
-          card_id: cardObj['Card ID'],
-          card_name: cardObj['Name'],
-          set_name: cardObj['Set Name'],
-          set_id: cardObj['Set ID'],
-          card_number: cardObj['Card Number'],
-          rarity: cardObj['Rarity'],
-          quantity: parseInt(cardObj['Quantity']) || 1,
-          condition: cardObj['Condition'] || 'Near Mint',
-          printing: cardObj['Printing'] || 'Normal',
-          language: cardObj['Language'] || 'English',
-          purchase_price: parseFloat(cardObj['Purchase Price']) || 0,
-          market_price: parseFloat(cardObj['Market Price']) || 0,
-          location_name: cardObj['Location Container'],
-          sub_location_1: cardObj['Sub-Location Page/Row'],
-          sub_location_2: cardObj['Sub-Location Slot/Section'],
-          added_at: cardObj['Added At']
-        });
-      }
-    }
-
-    if (!Array.isArray(cards)) {
-      return res.status(400).json({ error: 'Invalid data format. Expected an array or CSV lines.' });
-    }
-
-    let importedCount = 0;
-    for (const card of cards) {
-      const cardId = card.card_id || card.id;
-      if (!cardId) continue;
-
-      // 1. Ensure the card is in the cache. card_cache is shared across all users, so
-      // never trust client-supplied metadata for it beyond a sanitized last-resort placeholder.
-      let cached = await db.get(`SELECT id FROM card_cache WHERE id = ?`, [cardId]);
-      if (!cached) {
-        try {
-          const tcgApi = require('./tcgApi');
-          await tcgApi.getCardById(cardId);
-        } catch (apiErr) {
-          console.error(`Failed to fetch card ${cardId} from TCG API during import:`, apiErr.message);
-          const safeTypes = Array.isArray(card.types) ? JSON.stringify(card.types.filter(t => typeof t === 'string')) : '[]';
-          const safePrice = Number.isFinite(Number(card.market_price || card.price_trend)) ? Math.max(0, Number(card.market_price || card.price_trend)) : 0;
-          await db.run(
-            `INSERT OR IGNORE INTO card_cache
-             (id, name, supertype, subtypes, types, rarity, set_id, set_name, number, image_url, price_trend)
-             VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-            [
-              cardId,
-              String(card.card_name || 'Imported Card').slice(0, 200),
-              String(card.supertype || 'Pokémon').slice(0, 50),
-              '[]',
-              safeTypes,
-              String(card.rarity || 'Common').slice(0, 50),
-              String(card.set_id || '').slice(0, 50),
-              String(card.set_name || 'Imported Set').slice(0, 200),
-              String(card.card_number || card.number || '').slice(0, 20),
-              '',
-              safePrice
-            ]
-          );
-        }
-      }
-
-      // 2. Resolve location_id from location_name, scoped to this user only
-      let locationId = null;
-      const locName = card.location_name || card.location_container;
-      if (locName && locName !== 'Unassigned') {
-        let locRow = await db.get(`SELECT id FROM locations WHERE name = ? AND user_id = ?`, [locName, req.user.id]);
-        if (!locRow) {
-          let type = 'Other';
-          if (card.sub_location_1 && (card.sub_location_1.toLowerCase().includes('page') || card.sub_location_2)) {
-            type = 'Binder';
-          }
-          const newLoc = await db.run(`INSERT INTO locations (name, type, user_id) VALUES (?, ?, ?)`, [locName, type, req.user.id]);
-          locationId = newLoc.lastID;
-        } else {
-          locationId = locRow.id;
-        }
-      }
-
-      // 3. Insert card into the collection
-      await db.run(
-        `INSERT INTO collection 
-         (card_id, user_id, quantity, condition, printing, language, purchase_price, location_id, sub_location_1, sub_location_2, added_at)
-         VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
-        [
-          cardId,
-          req.user.id,
-          card.quantity || 1,
-          card.condition || 'Near Mint',
-          card.printing || 'Normal',
-          card.language || 'English',
-          card.purchase_price || 0,
-          locationId,
-          card.sub_location_1 || '',
-          card.sub_location_2 || '',
-          card.added_at || new Date().toISOString()
-        ]
-      );
-      importedCount++;
-    }
-
-    res.json({ success: true, message: `Successfully imported ${importedCount} cards.` });
-  } catch (error) {
-    console.error('Import failed:', error);
-    res.status(500).json({ error: 'Import failed' });
-  }
-});
-
-
-// --- ADVANCED COLLECTOR ENDPOINTS (DEX FEATURES) ---
-
-// 1. Get Card Price History
-app.get('/api/cards/:id/price-history', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  try {
-    let history = await db.all(`
-      SELECT price, recorded_at 
-      FROM price_history 
-      WHERE card_id = ? 
-      ORDER BY recorded_at ASC
-    `, [id]);
-
-    const cacheCard = await db.get(`SELECT price_trend FROM card_cache WHERE id = ?`, [id]);
-    const currentPrice = (cacheCard && cacheCard.price_trend) || 1.00;
-
-    // Seed mock price points if less than 5 records exist, so charts display immediately
-    if (history.length < 5) {
-      history = [];
-      const now = new Date();
-      for (let i = 9; i >= 0; i--) { // 10 price points
-        const date = new Date(now);
-        date.setDate(now.getDate() - i * 3); // 3 day intervals
-        const fluctuation = (Math.random() - 0.5) * 0.15 * currentPrice; // +/- 15%
-        const price = parseFloat(Math.max(0.10, currentPrice + fluctuation).toFixed(2));
-        history.push({
-          price,
-          recorded_at: date.toISOString()
-        });
-      }
-    } else {
-      // Map to standardized format
-      history = history.map(h => ({
-        price: h.price,
-        recorded_at: h.recorded_at
-      }));
-    }
-
-    res.json(history);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to retrieve price history' });
-  }
-});
-
-// 2. Get User Decks
-app.get('/api/decks', authenticateToken, async (req, res) => {
-  try {
-    const query = `
-      SELECT 
-        d.id,
-        d.name,
-        d.description,
-        d.created_at,
-        d.checked_out,
-        d.checked_out_at,
-        COUNT(dc.card_id) as total_card_types,
-        COALESCE(SUM(dc.quantity), 0) as total_cards
-      FROM decks d
-      LEFT JOIN deck_cards dc ON d.id = dc.deck_id
-      WHERE d.user_id = ?
-      GROUP BY d.id
-      ORDER BY d.created_at DESC
-    `;
-    const rows = await db.all(query, [req.user.id]);
-    res.json(rows);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to retrieve decks' });
-  }
-});
-
-// 3. Create Deck
-app.post('/api/decks', authenticateToken, async (req, res) => {
-  const { name, description = '' } = req.body;
-  if (!name) {
-    return res.status(400).json({ error: 'Deck name is required' });
-  }
-
-  try {
-    const result = await db.run(
-      `INSERT INTO decks (name, description, user_id) VALUES (?, ?, ?)`,
-      [name, description, req.user.id]
-    );
-    res.status(201).json({ message: 'Deck created successfully', id: result.lastID });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to create deck' });
-  }
-});
-
-// 4. Get Deck Details (with Cards)
-app.get('/api/decks/:id', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  try {
-    const deck = await db.get(`SELECT * FROM decks WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (!deck) {
-      return res.status(404).json({ error: 'Deck not found' });
-    }
-
-    const cardsQuery = `
-      SELECT 
-        dc.quantity,
-        cc.id,
-        cc.name,
-        cc.supertype,
-        cc.subtypes,
-        cc.types,
-        cc.rarity,
-        cc.set_id,
-        cc.set_name,
-        cc.number,
-        cc.image_url,
-        cc.price_trend
-      FROM deck_cards dc
-      JOIN card_cache cc ON dc.card_id = cc.id
-      WHERE dc.deck_id = ?
-    `;
-    const cards = await db.all(cardsQuery, [id]);
-
-    const formatted = cards.map(c => ({
-      ...c,
-      subtypes: JSON.parse(c.subtypes || '[]'),
-      types: JSON.parse(c.types || '[]')
-    }));
-
-    res.json({
-      ...deck,
-      cards: formatted
-    });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to retrieve deck details' });
-  }
-});
-
-// 5. Update Deck Metadata
-app.put('/api/decks/:id', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  const { name, description } = req.body;
-
-  if (!name) {
-    return res.status(400).json({ error: 'Deck name is required' });
-  }
-
-  try {
-    const result = await db.run(
-      `UPDATE decks SET name = ?, description = ? WHERE id = ? AND user_id = ?`,
-      [name, description || '', id, req.user.id]
-    );
-
-    if (result.changes === 0) {
-      return res.status(404).json({ error: 'Deck not found or unauthorized' });
-    }
-
-    res.json({ message: 'Deck updated successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to update deck' });
-  }
-});
-
-// 6. Delete Deck
-app.delete('/api/decks/:id', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  try {
-    // Verify ownership
-    const deck = await db.get(`SELECT id FROM decks WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (!deck) {
-      return res.status(404).json({ error: 'Deck not found or unauthorized' });
-    }
-
-    // Manual cascade deletion
-    await db.run(`DELETE FROM deck_cards WHERE deck_id = ?`, [id]);
-    await db.run(`DELETE FROM decks WHERE id = ?`, [id]);
-
-    res.json({ message: 'Deck deleted successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to delete deck' });
-  }
-});
-
-// 7. Add/Update Card in Deck
-app.post('/api/decks/:id/cards', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  const { card_id, quantity = 1 } = req.body;
-
-  if (!card_id) {
-    return res.status(400).json({ error: 'card_id is required' });
-  }
-
-  try {
-    // Verify deck ownership
-    const deck = await db.get(`SELECT id FROM decks WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (!deck) {
-      return res.status(404).json({ error: 'Deck not found or unauthorized' });
-    }
-
-    // Ensure card metadata exists in cache
-    let card = await db.get(`SELECT id FROM card_cache WHERE id = ?`, [card_id]);
-    if (!card) {
-      console.log(`Card ${card_id} not in cache. Fetching...`);
-      const apiCard = await tcgApi.getCardById(card_id, req.user.tcg_api_key);
-      if (!apiCard) {
-        return res.status(404).json({ error: 'Card not found on Pokémon TCG API.' });
-      }
-    }
-
-    // Insert or update quantities
-    await db.run(`
-      INSERT INTO deck_cards (deck_id, card_id, quantity)
-      VALUES (?, ?, ?)
-      ON CONFLICT(deck_id, card_id) DO UPDATE SET quantity = ?
-    `, [id, card_id, parseInt(quantity, 10), parseInt(quantity, 10)]);
-
-    // Record initial price history trend if card is added
-    const cacheCard = await db.get(`SELECT price_trend FROM card_cache WHERE id = ?`, [card_id]);
-    if (cacheCard && cacheCard.price_trend > 0) {
-      await db.run(`INSERT OR IGNORE INTO price_history (card_id, price) VALUES (?, ?)`, [card_id, cacheCard.price_trend]);
-    }
-
-    res.json({ message: 'Card added/updated in deck successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to add card to deck' });
-  }
-});
-
-// 8. Remove Card from Deck
-app.delete('/api/decks/:id/cards/:card_id', authenticateToken, async (req, res) => {
-  const { id, card_id } = req.params;
-  try {
-    // Verify deck ownership
-    const deck = await db.get(`SELECT id FROM decks WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (!deck) {
-      return res.status(404).json({ error: 'Deck not found or unauthorized' });
-    }
-
-    await db.run(`DELETE FROM deck_cards WHERE deck_id = ? AND card_id = ?`, [id, card_id]);
-    res.json({ message: 'Card removed from deck successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to remove card from deck' });
-  }
-});
-
-// 9. Checkout Deck (mark as in play)
-app.put('/api/decks/:id/checkout', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  try {
-    const deck = await db.get(`SELECT id FROM decks WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (!deck) {
-      return res.status(404).json({ error: 'Deck not found or unauthorized' });
-    }
-    await db.run(
-      `UPDATE decks SET checked_out = 1, checked_out_at = CURRENT_TIMESTAMP WHERE id = ?`,
-      [id]
-    );
-    res.json({ message: 'Deck checked out successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to checkout deck' });
-  }
-});
-
-// 10. Return Deck (mark as returned to storage)
-app.put('/api/decks/:id/return', authenticateToken, async (req, res) => {
-  const { id } = req.params;
-  try {
-    const deck = await db.get(`SELECT id FROM decks WHERE id = ? AND user_id = ?`, [id, req.user.id]);
-    if (!deck) {
-      return res.status(404).json({ error: 'Deck not found or unauthorized' });
-    }
-    await db.run(
-      `UPDATE decks SET checked_out = 0, checked_out_at = NULL WHERE id = ?`,
-      [id]
-    );
-    res.json({ message: 'Deck returned to storage successfully' });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: 'Failed to return deck' });
-  }
-});
-
+app.use('/api/auth', authRoutes);
+app.use('/api/shared', sharedRoutes);
+app.use('/api/admin', adminRoutes);
+app.use('/api', collectionRoutes);
+app.use('/api/decks', decksRoutes);
 
 // Serve production static assets from Frontend
 const frontendBuildPath = path.join(__dirname, '../../frontend/dist');
